requirements.txt
h5py==3.8.0
matplotlib==3.7.1
numpy==1.24.3
Pillow==9.5.0
scipy==1.10.1
